+++ java/dsbundle/src/main/java/com/dsbundle/BST/BinarySearchTree.java
@@ -1,4 +1,4 @@
-package com.dsbundle;
+package com.dsbundle.BST;
 
 import com.dsbundle.models.BSTNode;
